make authorisation

+++ app/Http/Controllers/PostController.php
@@ -10,6 +10,11 @@ use Illuminate\Support\Str;
 
 class PostController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except('index', 'show');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -56,7 +61,7 @@ class PostController extends Controller
         $post->title = $request->title;
         $post->short_title = Str::length($request->title) > 30 ? Str::substr($request->title, 0, 30) : $request->title;
         $post->descr = $request->descr;
-        $post->author_id = rand(1, 4);
+        $post->author_id = \Auth::user()->id;
 
         if ($request->file('img')) {
             $path = Storage::putFile('public', $request->file('img'));
@@ -79,6 +84,9 @@ class PostController extends Controller
     {
         $post = Post::join('users', 'author_id', '=', 'users.id')
             ->find($id);
+        if(!$post) {
+            return redirect()->route('post.index')->withErrors('Такого поста не существует');
+        }
         return view('posts.show', compact('post'));
     }
 
@@ -91,6 +99,12 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(!$post) {
+            return redirect()->route('post.index')->withErrors('Такого поста не существует');
+        }
+        if($post->author_id != \Auth::user()->id) {
+            return redirect()->route('post.index')->withErrors('Вы не можете редактировать данный пост');
+        }
         return view('posts.edit', compact('post'));
     }
 
@@ -104,6 +118,9 @@ class PostController extends Controller
     public function update(PostRequest $request, $id)
     {
         $post = Post::find($id);
+        if($post->author_id != \Auth::user()->id) {
+            return redirect()->route('post.index')->withErrors('Вы не можете редактировать данный пост');
+        }
         $post->title = $request->title;
         $post->short_title = Str::length($request->title) > 30 ? Str::substr($request->title, 0, 30) : $request->title;
         $post->descr = $request->descr;
@@ -130,6 +147,13 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(!$post) {
+            return redirect()->route('post.index')->withErrors('Такого поста не существует');
+        }
+        if($post->author_id != \Auth::user()->id) {
+            return redirect()->route('post.index')->withErrors('Вы не можете удалить данный пост');
+        }
+
         $post->delete();
         return redirect()->route('post.index')->with('success', 'Пост успешно удален');
     }
